backend/pyproject.toml
[project]
name = "schediora-backend"
version = "0.1.0"
description = "Schediora backend service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "email-validator>=2.2.0",
  "sqlalchemy>=2.0.36",
  "psycopg[binary]>=3.2.0",
  "alembic>=1.14.0",
  "pydantic-settings>=2.6.1",
  "python-jose[cryptography]>=3.3.0",
  "passlib[argon2]>=1.7.4",
  "redis>=5.2.0",
  "celery>=5.4.0",
  "httpx>=0.28.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.9.2",
  "mypy>=1.14.1",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
strict = false
